requirements.txt
PyQt5>=5.15.0
reportlab>=3.6.8
Pillow>=8.0.0 